requirements.txt
MetaTrader5==5.0.45
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
ta==0.10.2
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
tweepy==4.14.0
nltk==3.8.1
tensorflow==2.13.0
keras==2.13.1
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
python-telegram-bot==20.4
pymongo==4.5.0
pytest==7.4.2 